blog/requirements.txt
Flask-SQLAlchemy
wtforms_alchemy
pytest-flask
Flask-WTF
Flask-Mail
gunicorn